+++ CHANGELOG.md
@@ -1,5 +1,42 @@
-## [[tbd]] - 2023-08-04
-LT-5041: Update RabbitMqBroker NuGet package
+## 8.4.0 - 2024-09-26
+
+### Changed
+- LT-5719: bump Lykke.RabbitMqBroker -> 15.1.0
+
+## 8.3.0 - 2024-06-12
+
+### Added
+- LT-5509: Rabbit MQ listeners can be registered without autostart.
+
+## 8.2.0 - 2024-06-07
+
+### Added
+- LT-5509: RabbitMqBroker library version is now a part of connection display name
+
+## 8.1.1 - 2024-06-04
+
+### Fixed
+- LT-5509: Register dlx exchange and queue if configured when creating a new RabbitMQ subscriber from template.
+
+## 8.1.0 - 2024-05-30
+
+### Changed
+- LT-5509: Switch to template-based RabbiMQ subscribers
+
+## 8.0.2 - 2024-05-29
+
+### Changed
+- LT-5509: Objects disposal fixed in `RabbitMqPoisonQueueHandler`
+
+## 8.0.1 - 2024-05-29
+
+### Changed
+- LT-5509: Update RabbitMQ broker library with new RabbitMQ.Client and templates
+
+## 7.0.6 - 2023-08-04
+
+### Changed
+- LT-5041: Update RabbitMqBroker NuGet package
 
 ## 7.0.5 - 2023-08-04